feat(schema): align core user sector and qualification mapping

+++ dev/JobManagement/src/main/java/fr/imt_atlantique/fip/inf210/jobmanagement/repository/QualificationLevelRepository.java
@@ -1,10 +1,11 @@
 package fr.imt_atlantique.fip.inf210.jobmanagement.repository;
 
-import fr.imt_atlantique.fip.inf210.jobmanagement.entity.QualificationLevel;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import fr.imt_atlantique.fip.inf210.jobmanagement.entity.QualificationLevel;
+
 @Repository
-public interface QualificationLevelRepository extends JpaRepository<QualificationLevel, Long> {
+public interface QualificationLevelRepository extends JpaRepository<QualificationLevel, Integer> {
     
 }
