+++ dev/JobManagement/src/main/java/fr/imt_atlantique/fip/inf210/jobmanagement/repository/SectorJpaRepository.java
@@ -1,10 +1,11 @@
 package fr.imt_atlantique.fip.inf210.jobmanagement.repository;
 
-import fr.imt_atlantique.fip.inf210.jobmanagement.entity.Sector;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import fr.imt_atlantique.fip.inf210.jobmanagement.entity.Sector;
+
 @Repository
-public interface SectorJpaRepository extends JpaRepository<Sector, Long> {
+public interface SectorJpaRepository extends JpaRepository<Sector, Integer> {
     
 }
